Added support for testables in `Scheme`
Fixes #193

This addresses the issue that it was impossible to test targets if
custom scheme was used by providing array of testables to `Test` action.

Now it should be possible using the following syntax:

```yaml
STCore:
  build:
    targets:
      - target: STCore
        buildTypes: all
  test:
    testables: [STCoreTests]
    config: Staging-Debug
```

Note that original example in #193 has to be modified for this change
as it causes a confusion in Xcode: if we have 2 targets `STCore` and
`STCoreTests`, if `STCoreTests` is also added as testable Xcode
duplicates test target in build actions (i.e. `STCore`, `STCoreTests`,
`STCoreTests` are shown).