Rename ProjectSpec -> ProjectSpec.Project
I've started using the XcodeGen data structures in a large project, and
its been awesome so far.

The problem I'm running into is I can't import types from the
`ProjectSpec` module. This caused by a naming collision between the struct
`ProjectSpec` and the module `ProjectSpec`.

To remedy the issue, this patch renames the struct to `Project`. This is
an overall improvement, since all of the types are now nicely namespaced
and can be imported as expected.

i.e.
```
import ProjectSpec

let projectSpec: ProjectSpec.Project
let targetSpec: ProjectSpec.Target
```

This makes it easier to differentiate in the code when we are dealing
with types of similar names, like `PBXProject`