Added workaround for brew & Xcode 9.3 compatibility
`libxml2` has to be included in `ISYSTEM_PATH` for building one of
dependencies. It didn't happen automatically before Xcode 9.3
so `homebrew` patched environment variable to get it work.

But since Xcode 9.3 includes it already, the build will fail
because of redefinition of `libxml2` module.

It's a bug of `homebrew` but before it's fixed, it's easier
to provide in-place workaround for now.

Fixes #283