Fixed regression introduced in #289
I guess that's the reason for
https://github.com/yonaskolb/XcodeGen/issues/51#issuecomment-381216565,
I haven't modified original pull request after new discoveries found in
discussion #283 (primarily - that it is not Xcode 9.3 issue, but issue
caused by Xcode and CLT installed in the same time).

This patch modifies ENV only if CLT is installed.

I've checked it against Xcode + CLT and it worked fine, it'll be great
if someone can confirm that this one works on Xcode only system. You
can do that by running `brew edit xcodegen` and after patching
`brew reinstall xcodegen`