Make validation of Xcode projects optional
Validation requires several things of the Xcode project, one being that
files are already on disk. In some cases, it's useful to generate
projects containing files that don't exist on disk yet. Additionally,
this is useful for testing purposes, and generating projects on a
machine where the files don't exist.