Fix order of file generation
When the `Info.plist` doesn’t exist in filesystem (e.g. because it’s in `.gitignore`)
xcodegen generates the file but doesn’t add it to the `xcodeproj`. Only when
`xcodegen` is executed a second time it is included.

This changes the order of generation of these files to first generate the
plist and then the project.