Add better WidgetKit Extension support
This adds a new optional `subtype` property on `Target`. This is needed because Apple uses `app-extension` for WidgetKit Extensions, with only an entry in the Info.plist indicating which type of extension it is. By having a subtype declared, we are able to make a proper scheme for launching the extension, and provide proper defaults when using `info`.

This also adds the `hostTarget` property to `TargetScheme` and `Scheme.Run`. This allows specifying which target is hosting an extension or watch app. By specifying this it overrides the default logic in `TargetScheme` for selecting a host target to add to `Build` and it allows `SchemeGenerator` to add a `MacroExpansion` for WidgetKit Extensions.