Run preGenCommand before checking the cache (#1519)
The preGenCommand may generate files, change paths, etc. If we run this after checking the cache, it never gets the chance to run if the cache hasn't changed. Running it before checking the cache lets us make these changes and _then_ we can check the cache to see if we need to regenerate or not.

Co-authored-by: Dale Myers <dalemyers@microsoft.com>